ci: use MR project ID in commitlint script
Originally the script used `CI_PROJECT_ID` but that variable might
point to a forked project. The forked project is sometimes not the
target of the MR.

Now the `CI_MERGE_REQUEST_PROJECT_ID` always points to the project
with the MR `CI_MERGE_REQUEST_IID`.

See the following URL for CI variable documentation:
https://docs.gitlab.com/ee/ci/variables/predefined_variables.html