refactor: parseGitRemote returns object instead off an array
The only client of the parseGitRemote function is gitService and it took
the returned array and turned it into an object anyway. Returning object
allows us to annotate the parameters instead of relying on the index. We
will also be able to crate a return type once we migrate to TS