feat: suport detached pipelines
Prefer detached pipelines and only report normal branch pipeline when
there isn't any detached pipeline (MR pipeline).

We also change the logic to fetch MR and pipeline at the same time
because we need to know MR anyway to be able to get detached
pipeline.

Refactoring status_bar to use both pipeline and MR from the response
is too difficult to squeeze in this change. We'll have to do it later.