docs: don't prepend interfaces with capital I
This is a snippet from a book Effective TypeScript:

WARNING
The examples in this item prefix type names with I or T solely to
indicate how they were defined. You should not do this in your code!
Prefixing interface types with I is common in C#, and this convention
made some inroads in the early days of TypeScript. But it is considered
bad style today because it’s unnecessary, adds little value, and is not
consistently followed in the standard libraries.