refactor: introduce commands to start and cancel comment editing
Introduce startEdittingComment and cancelEdittingComment
commands that will open and close the VS Code comment
text area. The cancel command also resets the comment body
to the value retrieved from the API.

This is marked as refactor, because we don't want this commit to have
a line in the changelog. It's an uncomplete feature.