refactor: simplify the internal GraphQL response types for discussion
The real reason for this simplification is summed up in the following
line:

```
export type GqlNote = GqlTextDiffNote | GqlImageNote | GqlOverviewNote;
```

We can now refer to notes by GqlNote instead of having to specify
generic arguments. That makes the code easier to read and it will allow
us to return GqlNote from the createNote mutation.