refactor: simplify closing issue item
Instead of registering a new command, which was basically a callback, we
can declare what the item is supposed to do instead, reducing the amount
of procedural code.

Instead of triggering always one command which will decide what to do,
we can decide what to do at the time when we change the status bar item.