fix(status bar): hide all status items when there is no GitLab project
During the refactoring, we had to disable one test that hid the MR
status bar item when there wasn't any project. This commit fixes the
scenario by checking whether a GitLab project can be fetched for the
current workspace and if not, we hide all the status bar items. This
also happens to fix the #71.