refactor: expose repository in the TreeView items
This commit replaces the GitLabRepository(GitLabWorkspace) interface
with the Wrapped repository. The original objects `{ label, uri, error}`
were constructed by the issuable.ts.

We now use the WrappedRepository and the error state is signaled by a
new `containsGitLabProject` boolean flag. I.e. if the wrapped repository
doesn't contain GitLab project, there has been error during fetching it.

Also having access to the wrapped repository from within mr_item_model
is going to be important in the next commit because it allows us to
interact with the MR cache.