refactor: improve testability of WrappedRepository
This commit changes the way WrappedRepository exposes the
GitLabNewService. We used getter which was very hard to mock in tests.
Instead, we now use regular function (getGitLabService()) which is
easily replaced in tests.

Additionally, this commit creates a helper function for creating
WrappedRepository, because it is going to be used in many tests in the
future.